agents: split build/operate fleet (_dev/desk/shared), modes, G# guardrails
Restructure the agent layer by what each agent acts on:
- _dev/   — the build agent (code/tests/refactors); was agents/{skills,tools,memory}
- desk/   — operate agents (market/data/money), the default session mode; new
- shared/ — cross-agent substrate (knowledge -> MCP later); was agents/knowledge

Add vendor-neutral mode switching (default=desk, plain-text dev/desk signal,
session-start banner) in AGENTS.md, and a third requirement axis G# (desk
runtime guardrails) in agents/desk/GUARDRAILS.md alongside R#/D# — placed with
the agents it governs, not in docs/dev. Seed a reference agent
(desk/options-analyst: charter + pipeline). Bring TASKS.md back into the repo
at agents/_dev/TASKS.md as the dev agent's TODO cycle / learn-loop sink.

Fix recorder tool module paths (agents._dev.tools.exchange_fixtures) and every
cross-reference (AGENTS.md, README, PROJECT_OVERVIEW, DEVELOPMENT_REQUIREMENTS,
tests/utils, moved skills/memory). Record the decision in agents/_dev/memory.

Exchange suite stays hermetic: 38 passed, 3 integration deselected. The single
failure (test_get_provider_local) is the pre-existing missing-DATA_PATH gap
(T11), unrelated to this change.
