+++ agents/_dev/memory/agent-artifacts-layout.md
@@ -10,12 +10,20 @@ conflated two audiences and forced a split between a tool's *spec* (docs) and it
 Anthropic **Agent Skills** standard (a skill = a folder with its `SKILL.md` + `scripts/`).
 
 **Layout consequences:**
-- `agents/tools/<tool>.md` (spec) + `agents/tools/<tool_pkg>/` (code) sit together.
+- `agents/_dev/tools/<tool_pkg>/` (code) — the docstring is the doc; a `.md` spec is added
+  only for MCP/external tools whose config isn't in the code.
 - Test-only helpers are NOT agent tools → `tests/utils/` (imported by conftest).
 - `docs/dev/AI_WORKING_MODEL.md` was merged into `agents/README.md` (the working model
   now lives with the artifacts it describes).
 - Pytest `pythonpath` includes `.` so `agents.*` and `tests.utils.*` import as packages.
 
 Full rationale + compared practices (AGENTS.md standard, Agent Skills, repo conventions)
-in [agents/README.md](../README.md). This note graduates into AGENTS.md "Documentation
+in [agents/README.md](../../README.md). This note graduates into AGENTS.md "Documentation
 Conventions" (already updated) — kept here as the decision record.
+
+**Update 2026-06-15 — build/operate split.** `agents/` now separates the *build* agent
+(`_dev/`, underscore = private/special) from *operate*/desk agents (`desk/`) over a shared
+substrate (`shared/`, → MCP later). Default session mode is **desk**. See
+[agents/README.md](../../README.md) (modes + entity theses) and
+[agents/desk/GUARDRAILS.md](../../desk/GUARDRAILS.md) (G#). See also
+[agent-modes-and-layout.md](agent-modes-and-layout.md).
