+++ agents/_dev/memory/owner-verifies-math-and-architecture.md
