+++ agents/_dev/tools/exchange_fixtures/__init__.py
@@ -1,0 +1,5 @@
+"""Exchange fixture recording tool (AI tool, D4).
+
+Entry point: `python -m agents._dev.tools.exchange_fixtures` (see __main__.py).
+Indexed in agents/_dev/tools/README.md; playbook in agents/_dev/skills/refresh-exchange-fixtures.md.
+"""
