+++ agents/_dev/tools/exchange_fixtures/__main__.py
@@ -5,8 +5,8 @@ live venue APIs (needs network; run rarely — when an endpoint's shape changes)
 `alphavar.exchange` rather than re-implementing requests. Responses are keyed by
 path+query and stored with their HTTP status (so 4xx error paths are exercised too).
 
-    uv run python -m agents.tools.exchange_fixtures            # all exchanges
-    uv run python -m agents.tools.exchange_fixtures --only moex
+    uv run python -m agents._dev.tools.exchange_fixtures            # all exchanges
+    uv run python -m agents._dev.tools.exchange_fixtures --only moex
 
 After recording, shrink the fixtures (offline, idempotent) with the trimmer in
 tests/utils, then run the (now hermetic) exchange suite:
@@ -14,11 +14,11 @@ tests/utils, then run the (now hermetic) exchange suite:
     uv run python -m tests.utils.exchange_fixtures.trim
     uv run --extra etl pytest tests/unit/exchange -q
 
-Full playbook (when/why/verify): agents/skills/refresh-exchange-fixtures.md.
+Full playbook (when/why/verify): agents/_dev/skills/refresh-exchange-fixtures.md.
 """
 import argparse
 
-from agents.tools.exchange_fixtures import deribit, moex
+from agents._dev.tools.exchange_fixtures import deribit, moex
 
 _EXCHANGES = {'deribit': deribit.run, 'moex': moex.run}
