+++ agents/_dev/tools/exchange_fixtures/deribit.py
@@ -8,7 +8,7 @@ the bad call is skipped best-effort.
 from alphavar.exchange._abstract_exchange import RequestClass
 from alphavar.exchange.deribit import DeribitExchange, DeribitMarket, DeribitAssetKind
 
-from agents.tools.exchange_fixtures._record import record, try_call
+from agents._dev.tools.exchange_fixtures._record import record, try_call
 
 CURRENCY = 'BTC'
