+++ agents/desk/options-analyst/README.md
@@ -1,0 +1,25 @@
+# `options-analyst` (charter)
+
+**Role.** Find mispricings and notable structure in the options surface for a given
+underlying — **read-only** analysis, no trading.
+
+**Mode.** desk (default). Address directly ("as options-analyst …") or reach it via the
+orchestrator.
+
+**Scope.** Options/futures already supported by alphavar (Deribit, MOEX, …). In: an
+underlying `asset_code` + scope. Out: a sourced report (mispricing candidates, IV-surface
+notes), plus a `memory/` finding whenever something is missing or wrong.
+
+**Guardrails** ([`../GUARDRAILS.md`](../GUARDRAILS.md)): **G1** (every number sourced),
+**G2** (halt on stale/missing data), **G5** (read-only — never places an order).
+
+**Bound by R#** for the data model — uses the canonical column dictionary / pandera schemas
+through `alphavar`, never raw venue fields.
+
+**Pipeline.** [`pipeline.md`](pipeline.md).
+
+**Skills / tools / memory.** `skills/` and `tools/` are agent-local (promote to
+[`../../shared/`](../../shared/) on first reuse). `memory/findings.md` feeds the learn loop.
+
+**Success.** A reproducible, fully-sourced shortlist a human (or the orchestrator) can act
+on — every figure traceable to a quote timestamp + computation.
