+++ agents/shared/knowledge/options/README.md
