+++ agents/shared/knowledge/options/payoffs/README.md
