+++ agents/shared/knowledge/options/payoffs/single-leg.md
