+++ agents/shared/knowledge/options/pricing/greeks.md
