+++ agents/shared/knowledge/options/strategies/README.md
