+++ agents/shared/knowledge/options/strategies/straddle.md
