--- agents/tools/exchange_fixtures/__init__.py
@@ -1,5 +1,0 @@
-"""Exchange fixture recording tool (AI tool, D4).
-
-Entry point: `python -m agents.tools.exchange_fixtures` (see __main__.py).
-Indexed in agents/tools/README.md; playbook in agents/skills/refresh-exchange-fixtures.md.
-"""
