+++ tests/utils/exchange_fixtures/__init__.py
@@ -1,5 +1,5 @@
 """Exchange fixture test tooling: offline trimmer + httpx mock transport (T11).
 
-Replays the fixtures recorded by `agents/tools/exchange_fixtures`. Not an agent tool —
+Replays the fixtures recorded by `agents/_dev/tools/exchange_fixtures`. Not an agent tool —
 test infrastructure imported by tests/unit/exchange/conftest.py.
 """
